refactor : 댓글 달았을 때 실시간으로 반영하기 위한 파일 구조 개선 및 InputComment zustand 추가
