+++ src/app/(afterLogin)/posts/[postId]/page.tsx
@@ -1,6 +1,5 @@
 import DailyPostView from '@/app/(afterLogin)/posts/[postId]/_components/DailyPostView';
-import CommentView from '@/app/(afterLogin)/posts/[postId]/_components/comment/CommentView';
-import InputComment from '@/app/(afterLogin)/posts/[postId]/_components/input/InputComment';
+import CommentMain from '@/app/(afterLogin)/posts/[postId]/_components/comment/CommentMain';
 import Border from '@/app/_components/common/border/Border';
 import BasicHeader from '@/app/_components/common/header/BasicHeader';
 
@@ -19,11 +18,8 @@ export default function PostPage({ params }: Props) {
           <DailyPostView postId={params.postId} />
           <Border />
         </div>
-        <div className="flex flex-col flex-grow">
-          <CommentView postId={params.postId} />
-        </div>
+        <CommentMain postId={params.postId} />
       </div>
-      <InputComment postId={params.postId} />
     </>
   );
 }
